Fix u16 overflow in rate limit vs batch size validation
The comparison `(rate_limit as u16) < batch_size` caused u32 values
>= 65536 to wrap to 0, incorrectly rejecting valid rate limits.
Changed to `rate_limit < batch_size as u32` to compare in u32 space.
Also added --batch-size 1 to the property test so generated rate
limits (10-100000) don't fail the rate_limit >= batch_size check.

+++ src/config/args/mod.rs
@@ -675,7 +675,7 @@ impl TryFrom<CLIArgs> for Config {
 
         // Validate rate limit vs batch size
         if let Some(rate_limit) = args.rate_limit_objects {
-            if (rate_limit as u16) < batch_size {
+            if rate_limit < batch_size as u32 {
                 return Err(format!(
                     "--rate-limit-objects ({}) must be greater than or equal to --batch-size ({}).",
                     rate_limit, batch_size,
